Form Refactor + network type toolip
